+++ packages/reactor-api/src/types.ts
@@ -1,7 +1,7 @@
 import { type IAnalyticsStore } from "@powerhousedao/analytics-engine-core";
 import {
-  type IOperationalStore,
   type IProcessorManager,
+  type IRelationalDb,
 } from "document-drive/processors/types";
 import { type Express } from "express";
 import { type GraphQLManager } from "./graphql/graphql-manager.js";
@@ -9,7 +9,6 @@ import { type IPackageManager } from "./packages/types.js";
 
 export type { SubgraphClass } from "./graphql/index.js";
 export type { IPackageLoader } from "./packages/types.js";
-export type { Db } from "./utils/db.js";
 
 export type API = {
   app: Express;
@@ -20,5 +19,5 @@ export type API = {
 
 export type ReactorModule = {
   analyticsStore: IAnalyticsStore;
-  operationalStore: IOperationalStore;
+  relationalDb: IRelationalDb;
 };